Move OTB prefs to OTB directory
